rename lbrytv to web
