+++ .gitignore
@@ -13,5 +13,5 @@ package-lock.json
 .transifexrc
 .idea/
 /build/daemon*
-/lbrytv/dist/
-/lbrytv/node_modules
+/web/dist/*
+/web/node_modules/*
