+++ .lintstagedrc.json
@@ -1,9 +1,9 @@
 {
   "linters": {
     "ui/**/*.{js,jsx,scss,json}": ["prettier --write", "git add"],
-    "lbrytv/**/*.{js,jsx,scss,json}": ["prettier --write", "git add"],
+    "web/**/*.{js,jsx,scss,json}": ["prettier --write", "git add"],
     "ui/**/*.{js,jsx}": ["eslint", "flow focus-check --color always", "git add"],
-    "lbrytv/**/*.{js,jsx,scss}": ["eslint", "git add"]
+    "web/**/*.{js,jsx,scss}": ["eslint", "git add"]
   },
-  "ignore": ["node_modules", "dist", "package-lock.json"]
+  "ignore": ["node_modules", "web/dist/**/*", "dist/**/*", "package-lock.json"]
 }
