+++ ui/reducers.js
@@ -21,7 +21,7 @@ import {
   homepageReducer,
   statsReducer,
   syncReducer,
-  lbrytvReducer,
+  webReducer,
 } from 'lbryinc';
 import appReducer from 'redux/reducers/app';
 import contentReducer from 'redux/reducers/content';
@@ -53,5 +53,5 @@ export default history =>
     user: userReducer,
     wallet: walletReducer,
     sync: syncReducer,
-    lbrytv: lbrytvReducer,
+    web: webReducer,
   });
