+++ web/component/nag-degraded-performance.jsx
