+++ web/component/openInAppLink/index.js
