+++ web/component/openInAppLink/view.jsx
