+++ web/component/youtubeReferralWelcome/view.jsx
