+++ web/effects/use-degraded-performance.js
