+++ web/setup/publish.js
