+++ web/stubs/electron.js
